Review rating changed from Int to Float (#24)

+++ app/src/main/java/com/tue/yuni/models/review/CanteenReview.java
@@ -4,7 +4,7 @@ public class CanteenReview extends Review {
 
     private final int canteenId;
 
-    public CanteenReview(int id, int rating, String description, String createdAt, int canteenId) {
+    public CanteenReview(int id, float rating, String description, String createdAt, int canteenId) {
         super(id, rating, description, createdAt);
 
         this.canteenId = canteenId;
